add confusion matrix, mae, mse and boston dataset, modify kd tree solver, delete roaring map, modify sql scripts.

+++ drillbit/src/main/java/drillbit/parameter/Coordinates.java
@@ -17,7 +17,7 @@ public abstract class Coordinates {
         Dense, Sparse
     }
 
-    public double nDim() {
+    public int nDim() {
         return ndim;
     }
